AvatarPageTemplate used instead of EditUserForm
