--- webapp/src/shared/components/Input/EditUserForm.css
@@ -1,13 +1,0 @@
-.edit-user-form {
-  width: 100%;
-  color: var(--color-light);
-}
-
-.edit-user-form .inputs-container > .input {
-  margin-bottom: 0.875rem;
-}
-
-.edit-user-form .inputs-container > p:empty::before {
-  content: '';
-  display: inline-block;
-}
